Move fallback classes from HeapCodecFactory to the DirectCodecFactory
Remove HeapCodecFactory
